Spawn Loot Action

+++ Backend/Features/Loot/Interfaces/LootGenerationArgs.cs
@@ -4,6 +4,6 @@ namespace Mod.DynamicEncounters.Features.Loot.Interfaces;
 
 public class LootGenerationArgs
 {
-    public long MaxBudget { get; set; } = 1;
+    public double MaxBudget { get; set; } = 1;
     public IEnumerable<string> Tags { get; set; } = [];
 }
